Make alpaka/test headers installable
- move all test headers to the include folder
- test headers can be install via cmake option
- headers are not installed by default, because they are not designed
 for productive code
- built-in examples and tests are independent of the cmake option
