+++ include/alpaka/test/dim/TestDims.hpp
@@ -9,6 +9,7 @@
 
 #pragma once
 
+#include <alpaka/core/BoostPredef.hpp>
 #include <alpaka/dim/DimIntegralConst.hpp>
 
 #include <tuple>
