+++ include/alpaka/test/mem/view/ViewTest.hpp
